refactor(layout): 보호 레이아웃 이펙트 컴포넌트 분리 (#562)
